refactoring: modules renaming and restructuring
