+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "cim"
 description = "CDS Implementation generation by a Metadata document"
-version = "0.3.1"
+version = "0.3.2"
 edition = "2021"
 authors = ["zkud <anton.kudyrka@gmail.com>"]
 license = "MIT"
