+++ src/lib.rs
@@ -1,7 +1,8 @@
-mod parser;
+mod metadata;
+mod xml_tags;
 
 use clap::Parser;
-use parser::tag_parser::xml_rs::XmlTagParser;
+use xml_tags::xml_rs::XmlTagParser;
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -13,7 +14,7 @@ pub struct Args {
 pub fn run(args: Args) -> String {
   let path = args.path;
   let tag_parser = XmlTagParser::new(path);
-  let mut parser = parser::Parser::new(Box::new(tag_parser));
+  let mut parser = metadata::Parser::new(Box::new(tag_parser));
   let cds = parser.parse();
   cds
 }
