+++ src/metadata/util.rs
